Added more test classes with no implementation

+++ src/main/model/RedGiant.java
@@ -3,7 +3,7 @@ package model;
 public class RedGiant extends Giant {
 
     // EFFECT: construct a red giant based on luminosity
-    public RedGiant(int luminosity) {
+    public RedGiant(double luminosity) {
 
     }
